pokeapi-etl/requirements.txt
requests
sqlalchemy
psycopg2-binary
pandas
python-dotenv
pytest
pytest-mock
